<master> basic output to screen

+++ engine/src/color.h
@@ -6,9 +6,9 @@
 class Color : public Value
 {
   private:
-    unsigned int _rgba =0;
+    unsigned long int _rgba =0;
   public:
-  	explicit Color(unsigned int color) {_rgba = color;};
+  	explicit Color(unsigned long int color) {_rgba = color;};
   	Color() {Color(0);};
     void* asHSB(); //TODO pick a data format
     void* asHSL(); //TODO pick a data format
@@ -17,6 +17,6 @@ class Color : public Value
     void fromHSB(double h, double s, double b);
     void fromHSL(double h, double s, double l);
     void fromRGB(double r, double g, double b);
-    void fromRGBA(unsigned int rgba) {_rgba=rgba;};
+    void fromRGBA(unsigned long int rgba) {_rgba = rgba;};
 
 };
